+++ noodles-cram/src/writer/record.rs
@@ -454,7 +454,7 @@ where
         let preservation_map = self.compression_header.preservation_map();
         let tag_ids_dictionary = preservation_map.tag_ids_dictionary();
 
-        let keys: Vec<_> = record.tags().iter().map(|tag| tag.key()).collect();
+        let keys: Vec<_> = record.tags().values().map(|field| field.into()).collect();
         let tag_line = tag_ids_dictionary
             .iter()
             .enumerate()
@@ -471,17 +471,17 @@ where
 
         let tag_encoding_map = self.compression_header.tag_encoding_map();
 
-        for tag in record.tags() {
-            let id = tag.key().id();
-            let encoding = tag_encoding_map.get(&id).ok_or_else(|| {
+        for field in record.tags().values() {
+            let key: tag_ids_dictionary::Key = field.into();
+            let encoding = tag_encoding_map.get(&key.id()).ok_or_else(|| {
                 io::Error::new(
                     io::ErrorKind::InvalidInput,
-                    WriteRecordError::MissingTagEncoding(tag.key()),
+                    WriteRecordError::MissingTagEncoding(key),
                 )
             })?;
 
             let mut buf = Vec::new();
-            tag::write_value(&mut buf, tag.value())?;
+            tag::write_value(&mut buf, field.value())?;
 
             encode_byte_array(
                 encoding,